Adapt remaining packages to organization rename
