+++ client/src/main/java/de/upb/crypto/incentive/client/dto/GetEarnResponseDto.java
@@ -1,4 +1,4 @@
-package de.upb.crypto.incentive.client.dto;
+package org.cryptimeleon.incentivesystem.client.dto;
 
 
 import lombok.AllArgsConstructor;
