+++ client/src/test/java/de/upb/crypto/incentive/client/integrationtest/AliveTest.java
@@ -1,8 +1,8 @@
-package de.upb.crypto.incentive.client.integrationtest;
+package org.cryptimeleon.incentivesystem.client.integrationtest;
 
-import de.upb.crypto.incentive.client.BasketserverClient;
-import de.upb.crypto.incentive.client.CreditClient;
-import de.upb.crypto.incentive.client.IssueClient;
+import org.cryptimeleon.incentivesystem.client.BasketserverClient;
+import org.cryptimeleon.incentivesystem.client.CreditClient;
+import org.cryptimeleon.incentivesystem.client.IssueClient;
 import org.junit.jupiter.api.Test;
 import org.springframework.web.reactive.function.client.WebClient;
