--- credit/src/main/java/de/upb/crypto/incentive/services/credit/IncentiveException.java
@@ -1,11 +1,0 @@
-package de.upb.crypto.incentive.services.credit;
-
-import lombok.NoArgsConstructor;
-
-@NoArgsConstructor
-public class IncentiveException extends Exception {
-    IncentiveException(String msg) {
-        super(msg);
-    }
-
-}
