+++ credit/src/main/java/org/cryptimeleon/incentivesystem/services/credit/IncentiveException.java
@@ -1,4 +1,11 @@
 package org.cryptimeleon.incentivesystem.services.credit;
 
+import lombok.NoArgsConstructor;
+
+@NoArgsConstructor
 public class IncentiveException extends Exception {
+    IncentiveException(String msg) {
+        super(msg);
+    }
+
 }
