+++ credit/src/main/java/org/cryptimeleon/incentivesystem/services/credit/config/SwaggerConfig.java
@@ -1,4 +1,4 @@
-package de.upb.crypto.incentive.services.credit.config;
+package org.cryptimeleon.incentivesystem.services.credit.config;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
