+++ credit/src/main/java/org/cryptimeleon/incentivesystem/services/credit/mock/CryptoCreditMock.java
@@ -1,6 +1,6 @@
-package de.upb.crypto.incentive.services.credit.mock;
+package org.cryptimeleon.incentivesystem.services.credit.mock;
 
-import de.upb.crypto.incentive.cryptoprotocol.interfaces.provider.CreditInterface;
+import org.cryptimeleon.incentivesystem.cryptoprotocol.interfaces.provider.CreditInterface;
 
 /*
  * Mock for use until crypto credit-earn protocol is implemented
