+++ credit/src/test/java/org/cryptimeleon/incentivesystem/services/credit/CreditApplicationTests.java
@@ -1,4 +1,4 @@
-package de.upb.crypto.incentive.services.credit;
+package org.cryptimeleon.incentivesystem.services.credit;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
