+++ credit/src/test/java/org/cryptimeleon/incentivesystem/services/credit/TestConfiguration.java
@@ -1,0 +1,20 @@
+package org.cryptimeleon.incentivesystem.services.credit;
+
+import org.cryptimeleon.incentivesystem.cryptoprotocol.interfaces.provider.CreditInterface;
+import org.cryptimeleon.incentivesystem.services.credit.mock.TestBasketServerClientMock;
+import org.cryptimeleon.incentivesystem.services.credit.mock.TestCryptoCreditMock;
+import org.cryptimeleon.incentivesystem.services.credit.model.interfaces.BasketServerClientInterface;
+import org.springframework.context.annotation.Bean;
+
+public class TestConfiguration {
+
+    @Bean
+    BasketServerClientInterface basketServerClientInterface() {
+        return new TestBasketServerClientMock();
+    }
+
+    @Bean
+    CreditInterface creditInterface() {
+        return new TestCryptoCreditMock();
+    }
+}
