+++ credit/src/test/java/org/cryptimeleon/incentivesystem/services/credit/mock/TestCryptoCreditMock.java
@@ -1,6 +1,6 @@
-package de.upb.crypto.incentive.services.credit.mock;
+package org.cryptimeleon.incentivesystem.services.credit.mock;
 
-import de.upb.crypto.incentive.cryptoprotocol.interfaces.provider.CreditInterface;
+import org.cryptimeleon.incentivesystem.cryptoprotocol.interfaces.provider.CreditInterface;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
