+++ issue/src/main/java/org/cryptimeleon/incentivesystem/services/issue/IssueController.java
@@ -1,7 +1,7 @@
-package de.upb.crypto.incentive.services.issue;
+package org.cryptimeleon.incentivesystem.services.issue;
 
-import de.upb.crypto.incentive.services.issue.model.IssueResponse;
-import de.upb.crypto.incentive.services.issue.model.JoinRequest;
+import org.cryptimeleon.incentivesystem.services.issue.model.IssueResponse;
+import org.cryptimeleon.incentivesystem.services.issue.model.JoinRequest;
 import io.swagger.annotations.ApiOperation;
 import io.swagger.annotations.ApiResponse;
 import io.swagger.annotations.ApiResponses;
